Fix circular imports and incorrect use of typing.Optional
