+++ GCL/loss/__init__.py
@@ -1,15 +1,15 @@
 from .jsd import JSD, DebiasedJSD, HardnessJSD
 from .vicreg import VICReg
-from .infonce import InfoNCE, InfoNCESP, DebiasedInfoNCE, HardnessInfoNCE
+from .infonce import InfoNCE, DebiasedInfoNCE, HardnessInfoNCE
 from .triplet import TripletMargin, TripletMarginSP
 from .bootstrap import BootstrapLatent
 from .barlow_twins import BarlowTwins
-from .loss import Loss
+from .loss import Loss, ContrastInstance
 
 __all__ = [
     'Loss',
+    'ContrastInstance',
     'InfoNCE',
-    'InfoNCESP',
     'DebiasedInfoNCE',
     'HardnessInfoNCE',
     'JSD',
