+++ GCL/loss/barlow_twins.py
@@ -1,6 +1,6 @@
 import torch
 
-from GCL.loss import Loss
+from .loss import Loss
 
 
 def bt_loss(h1: torch.Tensor, h2: torch.Tensor, lambda_, batch_norm=True, eps=1e-15, *args, **kwargs):
